requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Validation & Settings
pydantic==2.9.2
pydantic-settings==2.4.0
email-validator==2.1.0.post1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv
python-multipart==1.0.0
phonenumbers==8.13.24

# Optimization
ortools==9.14.6206

# HTTP Client
httpx==0.25.0

# Cache (optional)
redis==5.0.1

# Development & Testing
pytest==7.4.3
aiohttp==3.9.1
